+++ News/Controllers/ArticleController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using News.Entities.Models;
+using News.Entities.Models.ModelsCreate;
 using News.Services.ServicesInterface;
 using System.Threading;
