+++ News/Controllers/PublisherController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using News.Entities.Models;
+using News.Entities.Models.ModelsCreate;
 using News.Services.ServicesInterface;
 
 namespace News.Controllers;
