+++ News/Entities/Models/ModelsCreate/ArticleCreateModel.cs
@@ -1,8 +1,8 @@
-﻿namespace News.Entities.Models;
+﻿namespace News.Entities.Models.ModelsCreate;
 
 public class ArticleCreateModel
-{  
-    public string? Title { get; set; }   
+{
+    public string? Title { get; set; }
     public string? SourceUrl { get; set; }
     public string? ImageUrl { get; set; }
     public string? Description { get; set; }
