+++ News/Entities/Models/ModelsCreate/CommentCreateModel.cs
@@ -1,4 +1,4 @@
-﻿namespace News.Entities.Models;
+﻿namespace News.Entities.Models.ModelsCreate;
 
 public class CommentCreateModel
 {
