+++ News/Entities/Models/ModelsCreate/TagCreateModel.cs
@@ -1,0 +1,6 @@
+﻿namespace News.Entities.Models.ModelsCreate;
+
+public class TagCreateModel
+{
+    public string? Name { get; set; }
+}
