+++ News/Entities/Models/ModelsCreate/UserCreateModel.cs
@@ -1,4 +1,4 @@
-﻿namespace News.Entities.Models;
+﻿namespace News.Entities.Models.ModelsCreate;
 
 public class UserCreateModel
 {
