+++ News/Entities/Models/ModelsUpdate/UserUpdateModel.cs
@@ -1,9 +1,7 @@
-﻿namespace News.Entities.Models;
+﻿namespace News.Entities.Models.ModelsUpdate;
 
-public class UserResponseModel
+public class UserUpdateModel
 {
-    public Guid? Id { get; set; }
-
     public string? FirstName { get; set; }
 
     public string? Email { get; set; }
