--- News/Entities/Models/TagCreateModel.cs
@@ -1,6 +1,0 @@
-﻿namespace News.Entities.Models;
-
-public class TagCreateModel
-{
-    public string? Name { get; set; }    
-}
