+++ News/Infrastructure/InfrastructureInterface/IPasswordEncryptionHelper.cs
