+++ News/Infrastructure/InfrastructureInterface/ITokenHelper.cs
@@ -1,4 +1,4 @@
-﻿namespace News.Infrastructure;
+﻿namespace News.Infrastructure.IInfrastructure;
 
 public interface ITokenHelper
 {
