+++ News/Infrastructure/PasswordEncryptionHelper.cs
@@ -1,4 +1,5 @@
 ﻿using System.Security.Cryptography;
+using News.Infrastructure.IInfrastructure;
 
 namespace News.Infrastructure;
