+++ News/Mapping/NewsMappingProfile.cs
@@ -1,6 +1,9 @@
 ﻿using AutoMapper;
 using News.Entities;
 using News.Entities.Models;
+using News.Entities.Models.ModelsCreate;
+using News.Entities.Models.ModelsRespones;
+using News.Entities.Models.ModelsUpdate;
 
 namespace News.Mapping;
