+++ News/Services/ArticleService.cs
@@ -2,6 +2,7 @@
 using News.DataAccess.Repo;
 using News.Entities;
 using News.Entities.Models;
+using News.Entities.Models.ModelsCreate;
 using News.Services.ServicesInterface;
 using System.IO;
