+++ News/Services/CommentService.cs
@@ -2,6 +2,10 @@
 using News.DataAccess.Repo.RepoInterfaces;
 using News.Entities;
 using News.Entities.Models;
+using News.Entities.Models.ModelsCreate;
+using News.Entities.Models.ModelsRespones;
+using News.Entities.Models.ModelsUpdate;
+using News.Services.ServicesInterface;
 
 namespace News.Services;
