+++ News/Services/PublisherService.cs
@@ -4,6 +4,7 @@ using News.DataAccess.Repo;
 using News.DataAccess.Repo.RepoInterfaces;
 using News.Entities;
 using News.Entities.Models;
+using News.Entities.Models.ModelsCreate;
 using News.Services.ServicesInterface;
 
 namespace News.Services
