+++ News/Services/ServicesInterface/ICommentService.cs
@@ -1,6 +1,9 @@
 ﻿using News.Entities.Models;
+using News.Entities.Models.ModelsCreate;
+using News.Entities.Models.ModelsRespones;
+using News.Entities.Models.ModelsUpdate;
 
-namespace News.Services;
+namespace News.Services.ServicesInterface;
 
 public interface ICommentService
 {
