+++ News/Services/TokenService.cs
@@ -3,7 +3,10 @@ using Microsoft.IdentityModel.Tokens;
 using News.DataAccess.Repo.RepoInterfaces;
 using News.Entities;
 using News.Entities.Models;
+using News.Entities.Models.ModelsRespones;
 using News.Infrastructure;
+using News.Infrastructure.IInfrastructure;
+using News.Services.ServicesInterface;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Security.Cryptography;
